requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
passlib==1.7.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.0.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
